fix: resolve narrowing conversion warnings (bugprone-narrowing-conversions)
Applied safety fixes to prevent implicit type conversions that could cause
data loss or unexpected behavior:

- manager.cc: Use double for atof() result (line 2969), size_t for strlen
  results (lines 3088-3089) instead of int
- zone.cc: Add explicit static_cast<short> for Page.size assignment (1274),
  static_cast<int> for strlen (1496)
- user_edit_zone.cc: Add explicit static_cast<short> for FormField::Set()
  call with int parameter (739)

These changes follow clang-tidy bugprone-narrowing-conversions recommendations
and improve type safety across the codebase.

Verification: Build successful, all 40 tests passing.
Updated changelog.md to document these safety improvements.

+++ zone/user_edit_zone.cc
@@ -733,7 +733,7 @@ int JobSecurityZone::LoadRecord(Terminal *term, int record)
         f = f->next;
 
         // Active switch
-        f->Set(a); f = f->next;
+        f->Set(static_cast<short>(a)); f = f->next;
 
         // tables
         f->active = a;
